+++ src/components/main/buttons/index.js
@@ -1,2 +1,3 @@
 export { default as ChattingBtn } from './ChattingBtn';
 export { default as EditBtn } from './EditBtn';
+export { default as UploadBtn } from './UploadBtn';
