+++ src/components/main/index.js
@@ -1,4 +1,3 @@
 export { default as Description } from './Description';
 export { default as ExampleQuestion } from './ExampleQuestion';
-export { default as UploadBtn } from './UploadBtn';
 export { default as MainContents } from './MainContents';
